refactor(interface): rename ExecResult → CommandResult, code: i32
- ExecResult → CommandResult (clearer name for command output)
- code: u8 → code: i32 (POSIX exit codes can exceed 255)
- ABI Module struct renamed from Command to Module
  (avoids conflict with rush_core::types::Command)
- All plugins and macros updated to new types

Breaking: existing .so plugins must be rebuilt.

56/56 tests pass, zero clippy warnings.

+++ rush-core/src/var.rs
@@ -36,7 +36,7 @@ impl VarStore {
     }
 
     /// Update the `$?` exit code.
-    pub fn set_exit_code(&self, code: u8) {
+    pub fn set_exit_code(&self, code: i32) {
         self.set("?", vec![code.to_string()]);
     }
